Return raw bytes (#4)
Return the position of dictionaries in the raw byte stream. They are placed under a custom key in the dictionary.

This allows to overwrite a project with only the modified parts. Xcode projects contain comments and are sensitive to ordering of objects, this allows to be more compatible with Cordova and the Xcode agvtool tool.